Renamed `Dev/` to `headers/` and `Structs/` to `structs/`
