+++ README.md
@@ -1,19 +1,25 @@
 # Library for the C language
 
-Includes:
-* Structs
-  - List (dynamic array)
-  - String (dynamic string)
-  - Set (hash table)
-  - Dictionary (hash map)
-* Headers (`Dev/`) for:
-  - Unit testing & assertions
-  - Error handling
-  - Terminal colors (for UNIX systems)
-  - ...
-* Misc:
-  - Some mathematical functions
-  - String utilities
-  - Switch "expressions"
-  - ...
+---
 
+## Includes
+
+- List (dynamic array)
+- String (dynamic string)
+- Set (hash table)
+- Dictionary (hash map)
+- Unit testing & assertions
+- Error handling
+- Terminal colors (for UNIX systems)
+- Some mathematical functions
+- String utilities
+- Switch "expressions"
+- ...
+
+---
+
+## Structure
+
+* `src/headers/` – (mostly) lightweight headers without the need for linking/…
+* `src/structs/` – advanced data structures
+* `src/*.[ch]` – useful algorithms
